requirements.txt
streamlit
deepface
opencv-python
numpy
pillow